rename rc-loader to spashiprc-loader
