+++ packages/cli/src/common/spashiprc-loader.js
